servers/plane-mcp/pyproject.toml
[project]
name = "plane-mcp"
version = "0.1.0"
description = "Governance-safe MCP wrapper for Plane project management"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "plane-sdk>=0.2.0,<0.3.0",
    "pydantic>=2.5.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "asyncpg>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
